feat: unir delete use case
